[fix] linkItemをconstant化
