+++ view/next-project/src/components/common/SideNav.tsx
@@ -1,17 +1,7 @@
 import clsx from 'clsx';
 import { useRouter } from 'next/router';
 import React, { ReactNode, useReducer } from 'react';
-import { BsBuilding, BsVectorPen } from 'react-icons/bs';
-import { HiOutlineShoppingCart, HiOutlineDocumentText, HiCurrencyDollar } from 'react-icons/hi';
-import { IoIosArrowDropup } from 'react-icons/io';
-import { MdOutlineSavings, MdOutlineWorkOutline } from 'react-icons/md';
-
-interface LinkItemProps {
-  name: string;
-  icon: ReactNode;
-  href: string;
-  isParent?: boolean;
-}
+import { FinanceLinkItems, RelationLinkItems } from '@/constants/linkItem';
 
 interface NavItemProps {
   icon: ReactNode;
@@ -22,63 +12,6 @@ interface NavItemProps {
   isShow?: boolean;
   onClick?: () => void;
 }
-
-const FinanceLinkItems: LinkItemProps[] = [
-  {
-    name: '財務局',
-    icon: <IoIosArrowDropup className='mx-2 text-xl' />,
-    href: '',
-    isParent: true,
-  },
-  {
-    name: '予算',
-    icon: <HiCurrencyDollar className='mx-2 text-xl' />,
-    href: '/budgets',
-  },
-  {
-    name: '学内募金',
-    icon: <MdOutlineSavings className='mx-2 text-xl' />,
-    href: '/fund_informations',
-  },
-  {
-    name: '購入申請',
-    icon: <HiOutlineShoppingCart className='mx-2 text-xl' />,
-    href: '/purchaseorders',
-  },
-  {
-    name: '購入報告',
-    icon: <HiOutlineDocumentText className='mx-2 text-xl' />,
-    href: '/purchasereports',
-  },
-];
-
-const RelationLinkItems: LinkItemProps[] = [
-  {
-    name: '渉外局',
-    icon: <IoIosArrowDropup className='mx-2 text-xl' />,
-    href: '',
-    isParent: true,
-  },
-  {
-    name: '協賛活動',
-    icon: <MdOutlineWorkOutline className='mx-2 text-xl' />,
-    href: '/sponsoractivities',
-  },
-  {
-    name: '協賛スタイル',
-    icon: <BsVectorPen className='mx-2 text-xl' />,
-    href: '/sponsorstyles',
-  },
-  {
-    name: '協賛企業一覧',
-    icon: <BsBuilding className='mx-2 text-xl' />,
-    href: '/sponsors',
-  },
-];
-
-const otherLinkItems: LinkItemProps[] = [
-  // { name: 'ダッシュボード', icon: MdOutlineDashboard, href: '/' },
-];
 
 export default function SimpleSidebar() {
   const router = useRouter();
